DocuMind AI — a Multimodal RAG Document Intelligence System/requirement.txt
flask
python-dotenv
sentence-transformers
faiss-cpu
numpy
pymupdf
pytesseract
pillow
langchain-text-splitters
groq
werkzeug